refactor(php): move MPP-only PayCore files into Protocols/Mpp/Core
Per Ludo's placement rule (MPP-specific -> Protocols/Mpp,
x402-specific -> Protocols/X402, shared by both -> PayCore,
building on top -> umbrella root), the following PayCore files are
in fact only used by Protocols/Mpp/* and move accordingly:

- src/PayCore/Base64Url.php       -> src/Protocols/Mpp/Core/Base64Url.php
- src/PayCore/Challenge.php       -> src/Protocols/Mpp/Core/Challenge.php
- src/PayCore/ChallengeEcho.php   -> src/Protocols/Mpp/Core/ChallengeEcho.php
- src/PayCore/Credential.php      -> src/Protocols/Mpp/Core/Credential.php
- src/PayCore/Headers.php         -> src/Protocols/Mpp/Core/Headers.php
- src/PayCore/Json.php            -> src/Protocols/Mpp/Core/Json.php
- src/PayCore/Receipt.php         -> src/Protocols/Mpp/Core/Receipt.php
- src/PayCore/Rfc3339Parser.php   -> src/Protocols/Mpp/Core/Rfc3339Parser.php

Cross-checked: x402 uses native base64_* (not Base64Url), has its
own typed PaymentRequiredEnvelope (no Headers / Json overlap with
MPP), and doesn't surface RFC 3339 expiries (uses
maxTimeoutSeconds). The directory shape now mirrors ruby/lib at
lib/mpp/protocol/core/{base64url,challenge,challenge_echo,credential,
receipt,headers,json,rfc3339_parser}.

Stays in src/PayCore/ (genuinely shared by both protocols + the
umbrella):
- Currency.php / Network.php / Stablecoin.php (typed value enums)
- HttpFactory.php (PSR-17 factories + SAPI emit used by Middleware)
- Solana/Mints.php (ticker -> mint pubkey resolver, used by both
  X402 Adapter::acceptsEntry and MPP charge verifier)

Test mirrors move alongside: tests/PayCore/{Base64Url,Challenge,
ChallengeEcho,Credential,Headers,Json,Receipt,Rfc3339Parser}Test.php
-> tests/Protocols/Mpp/Core/. tests/PayCore/ keeps HttpFactoryTest
+ MintsTest (the genuinely shared bits).

PHP suite: 315 tests / 636 assertions / 0 failures. Lint, validate,
audit all clean. Coverage 91.47% (gate 91). Manual smoke against
the simple-server example confirms 402 with both x402 + mpp accept
entries.
