+++ php/tests/Protocols/Mpp/Core/ReceiptTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace PayKit\Tests;
+namespace PayKit\Tests\Protocols\Mpp\Core;
 
 use DateTimeImmutable;
 use DateTimeZone;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
-use PayKit\PayCore\Receipt;
+use PayKit\Protocols\Mpp\Core\Receipt;
 
 final class ReceiptTest extends TestCase
 {
