+++ php/tests/Protocols/Mpp/Server/ChargeServerTest.php
@@ -7,10 +7,10 @@ namespace PayKit\Tests;
 use DateTimeImmutable;
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
-use PayKit\PayCore\Challenge;
-use PayKit\PayCore\Credential;
-use PayKit\PayCore\Headers;
-use PayKit\PayCore\Json;
+use PayKit\Protocols\Mpp\Core\Challenge;
+use PayKit\Protocols\Mpp\Core\Credential;
+use PayKit\Protocols\Mpp\Core\Headers;
+use PayKit\Protocols\Mpp\Core\Json;
 use PayKit\Protocols\Mpp\Intent\ChargeRequest;
 use PayKit\Protocols\Mpp\Server\ChargeServer;
 use PayKit\Protocols\Mpp\Server\PaymentVerifier;
