+++ php/tests/Protocols/Mpp/Server/SolanaChargeHandlerTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace PayKit\Tests;
 
 use PHPUnit\Framework\TestCase;
-use PayKit\PayCore\Challenge;
-use PayKit\PayCore\Credential;
+use PayKit\Protocols\Mpp\Core\Challenge;
+use PayKit\Protocols\Mpp\Core\Credential;
 use PayKit\Protocols\Mpp\Intent\ChargeRequest;
 use PayKit\Protocols\Mpp\Server\ChargeServer;
 use PayKit\Protocols\Mpp\Server\ChargeSettlement;
